requirements/doc.txt
-r base.txt
sphinx~=7.2
sphinx_rtd_theme~=1.2
autodoc_pydantic~=2.0